General refactor of routes Added link to project from add technology to project page
