+++ routes/web/project-routes.js
@@ -1,5 +1,4 @@
-var projects = require('../../dao/projects');
-var technology = require('../../dao/technology');
+var handler = require('../../handlers/web/projectsWebHandler');
 var security = require('../../utils/security.js');
 
 var ProjectRoutes = function () {
@@ -11,60 +10,22 @@ ProjectRoutes.createRoutes = function (self) {
     /**
      * Add new project page
      */
-    self.app.get('/project/add', security.isAuthenticatedAdmin,
-        function (req, res) {
-            res.render('pages/admin/addProject', {user: req.user});
-        });
-
-
+    self.app.get('/project/add', security.isAuthenticatedAdmin, handler.add);
+    
     /**
      * Edit project page
      */
-    self.app.get('/project/:projectId/edit', security.isAuthenticatedAdmin,
-        function (req, res) {
-            projects.findById(req.params.projectId, function (error, project) {
-                if (error) {
-                    res.render('pages/error', {user: req.user});
-                } else {
-                    res.render('pages/admin/editProject', {user: req.user, project: project});
-                }
-            });
-
-        });
+    self.app.get('/project/:projectId/edit', security.isAuthenticatedAdmin, handler.edit);
 
     /**
      * Show project radar page
      */
-    self.app.get('/project/:projectId', security.isAuthenticated,
-        function (req, res) {
-            projects.findById(req.params.projectId, function (error, project) {
-
-                if (error) {
-                    res.render('pages/error', {user: req.user});
-                } else {
-                    technology.getAllForProject(project.id, function (error, technologies) {
-                        if (error) {
-                            res.render('pages/error', {user: req.user});
-                        } else {
-                            res.render('pages/projectRadar', {
-                                user: req.user,
-                                project: project,
-                                technologies: technologies
-                            });
-                        }
-                    });
-                }
-            });
-
-        });
+    self.app.get('/project/:projectId', security.isAuthenticated, handler.showRadar);
 
     /**
      * List projects page
      */
-    self.app.get('/projects', security.isAuthenticated,
-        function (req, res) {
-            res.render('pages/listProjects', {user: req.user});
-        });
+    self.app.get('/projects', security.isAuthenticated, handler.list );
 
 
 }
