+++ routes/web/routes.js
@@ -1,12 +1,9 @@
 /**
- *  These are the routes for the Web Application
+ *  Main routes for the Application
  */
-
 var comments = require('../../dao/comments.js');
 var technology = require('../../dao/technology.js');
-
 var passport = require('passport');
-
 var security = require('../../utils/security.js');
 
 
@@ -68,11 +65,6 @@ Routes.createRoutes = function (self) {
         req.logout();
         res.redirect('/login');
     });
-
-    
-
-  
-  
 
 }
