+++ routes/web/user-routes.js
@@ -1,5 +1,5 @@
-var users = require('../../dao/users');
-var security = require('../../utils/security.js');
+var handler = require('../../handlers/web/usersWebHandler');
+var security = require('../../utils/security');
 
 var UserRoutes = function () {
 };
@@ -10,42 +10,23 @@ UserRoutes.createRoutes = function (self) {
     /**
      * Users page
      */
-    self.app.get('/users', security.isAuthenticatedAdmin,
-        function (req, res) {
-            res.render('pages/admin/user/listUsers', {user: req.user});
-        });
+    self.app.get('/users', security.isAuthenticatedAdmin, handler.list);
 
 
     /**
      * Add new user page
      */
-    self.app.get('/user/add', security.isAuthenticatedAdmin,
-        function (req, res) {
-            res.render('pages/admin/user/addUser', {user: req.user});
-        });
+    self.app.get('/user/add', security.isAuthenticatedAdmin, handler.add );
 
     /**
-     * Edit profile page
+     * Edit the current users profile page
      */
-    self.app.get('/profile', security.isAuthenticated,
-        function (req, res) {
-            res.render('pages/editProfile', {user: req.user, editUser: req.user});
-        });
+    self.app.get('/profile', security.isAuthenticated, handler.editProfile );
 
     /**
-     * Edit user page
+     * Edit specific user page
      */
-    self.app.get('/user/:userId/edit', security.isAuthenticatedAdmin,
-        function (req, res) {
-            users.findById(req.params.userId, function (error, editUser) {
-                if (error) {
-                    res.render('pages/error', {user: req.user});
-                } else {
-                    res.render('pages/admin/user/editUser', {user: req.user, editUser: editUser});
-                }
-            });
-
-        });
+    self.app.get('/user/:userId/edit', security.isAuthenticatedAdmin, handler.editUser );
 
 }
